src/server/requirements.txt
requests
flask
uwsgi
neo4j-driver
globus-sdk
celery
redis
pycrypto
